+++ R/initialize_modules.R
@@ -856,8 +856,9 @@ initialize_comp <- function(data,
 #'   fleet exists in the data but parameter information for how to specify
 #'   selectivity for that fleet is not provided, then selectivity will not be
 #'   initialized for that fleet.
-#' @param data An S4 object with the `FIMSFrame` class, which is returned from
-#'   [FIMSFrame()]. Passing the data is required because initialization of the
+#' @param data A `FIMSFrame` object or a data frame accepted by [FIMSFrame()].
+#'   Data frames are converted before initialization. Passing the data is
+#'   required because initialization of the
 #'   modules requires passing the data and information regarding the uncertainty
 #'   of that data, i.e., input sample sizes for the multinomial distribution.
 #' @param recruitment_schedule NULL for annual January 1 recruitment, or a
@@ -918,6 +919,8 @@ initialize_fims <- function(parameters, data, recruitment_schedule = NULL) {
     ))
   }
 
+  # Normalize table inputs before validating schedules and observation timing.
+  if (!methods::is(data, "FIMSFrame")) data <- FIMSFrame(data)
   phases <- setup_recruitment_schedule(data, recruitment_schedule)
 
   samples <- observation_table(data)